投稿の編集ボタン、編集画面実装

+++ post/post_edit_done.php
@@ -1,55 +1,64 @@
-<?php require_once('../head.php'); ?>
-<body>
-
-<?php
+<?php 
+require_once('../config.php'); 
 
 try
 {
-    $post_id = $_POST['id'];
-    $post_name = $_POST['name'];
-    $post_price = $_POST['price'];
-    $post_gazou_name_old = $_POST['gazou_name_old'];
-    $post_gazou_name = $_POST['gazou_name'];
+$post_id = $_POST['id'];
+$post_text = $_POST['text'];
+$post_gazou_name_old = $_POST['gazou_name_old'];
+$post_gazou_name = $_FILES['gazou_name'];
 
-    $post_id=htmlspecialchars($post_id,ENT_QUOTES,'UTF-8');
-    $post_name=htmlspecialchars($post_name,ENT_QUOTES,'UTF-8');
-    $post_price=htmlspecialchars($post_price,ENT_QUOTES,'UTF-8');
-
-    $dsn = 'mysql:dbname=shop;host=localhost';
-    $user = 'root';
-    $password = '';
-    $dbh = new PDO($dsn,$user,$password);
-    $dbh -> setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-
-    $sql = 'UPDATE post SET name=?,price=?,gazou=? WHERE id=?';
-    $stmt = $dbh -> prepare($sql);
-    $data[] = $post_name;
-    $data[] = $post_price;
-    $data[] = $post_gazou_name;
-    $data[] = $post_id;
-    $stmt -> execute($data);
-
-    $dbh = null;
-
-if($post_gazou_name_old!=$post_gazou_name)
+if($post_text=='')
 {
-    if($post_gazou_name_old!='')
+    set_flash('danger','投稿内容が未記入です');
+    reload();
+}
+
+if($post_gazou_name['size']>0)
+{
+    if($post_gazou_name['size']>1000000)
     {
-        unlink('./gazou/'.$post_gazou_name_old);
+        set_flash('danger','画像が大きすぎます');
+        reload();
+    }
+    else
+    {
+        move_uploaded_file($post_gazou_name['tmp_name'],'./gazou/'.$post_gazou_name['name']);
+
     }
 }
-    print '修正しました。<br />';
 
-}   
+$post_text=htmlspecialchars($post_text,ENT_QUOTES,'UTF-8');
+$post_id=htmlspecialchars($post_id,ENT_QUOTES,'UTF-8');
+
+$dsn = 'mysql:dbname=shop;host=localhost';
+$user = 'root';
+$password = '';
+$dbh = new PDO($dsn,$user,$password);
+$dbh -> setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+
+$sql = 'UPDATE post SET text=?,gazou=? WHERE id=?';
+$stmt = $dbh -> prepare($sql);
+$data[] = $post_text;
+$data[] = $post_gazou_name['name'];
+$data[] = $post_id;
+$stmt -> execute($data);
+
+$dbh = null;
+
+if($post_gazou_name_old!='' && $post_gazou_name_old!=$post_gazou_name['name'])
+{
+    unlink('./gazou/'.$post_gazou_name_old);
+}
+}
 catch (Exception $e)
 {
     print'ただいま障害により大変ご迷惑をお掛けしております。';
     exit();
 }
 
-?>
+set_flash('sucsess','更新しました');
+reload();
 
-<a href="post_index.php">戻る</a>
-</body>
-<?php require_once('../footer.php'); ?>
-</html>
+require_once('../footer.php');
+?>
